Make demo arguments parsing platform-independent
