+++ demos/app/iot_demo.c
@@ -83,133 +83,6 @@ extern int RunDemo( bool awsIotMqttMode,
 
 /*-----------------------------------------------------------*/
 
-/**
- * @brief Set the default values of an #IotDemoArguments_t based on compile-time
- * defined constants.
- *
- * @param[out] pArguments Default values will be placed here.
- */
-static void _setDefaultArguments( IotDemoArguments_t * pArguments )
-{
-    /* Default to AWS IoT MQTT mode. */
-    pArguments->awsIotMqttMode = true;
-
-    /* Set default identifier if defined. The identifier is used as either the
-     * MQTT client identifier or the Thing Name, which identifies this client to
-     * the cloud. */
-    #ifdef IOT_DEMO_IDENTIFIER
-        pArguments->pIdentifier = IOT_DEMO_IDENTIFIER;
-    #endif
-
-    /* Set default secured connection status if defined. */
-    #ifdef IOT_DEMO_SECURED_CONNECTION
-        pArguments->securedConnection = IOT_DEMO_SECURED_CONNECTION;
-    #endif
-
-    /* Set default MQTT server if defined. */
-    #ifdef IOT_DEMO_SERVER
-        pArguments->pHostName = IOT_DEMO_SERVER;
-    #endif
-
-    /* Set default MQTT server port if defined. */
-    #ifdef IOT_DEMO_PORT
-        pArguments->port = IOT_DEMO_PORT;
-    #endif
-
-    /* Set default root CA path if defined. */
-    #ifdef IOT_DEMO_ROOT_CA
-        pArguments->pRootCaPath = IOT_DEMO_ROOT_CA;
-    #endif
-
-    /* Set default client certificate path if defined. */
-    #ifdef IOT_DEMO_CLIENT_CERT
-        pArguments->pClientCertPath = IOT_DEMO_CLIENT_CERT;
-    #endif
-
-    /* Set default client certificate private key path if defined. */
-    #ifdef IOT_DEMO_PRIVATE_KEY
-        pArguments->pPrivateKeyPath = IOT_DEMO_PRIVATE_KEY;
-    #endif
-}
-
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Validates the members of an #IotDemoArguments_t.
- *
- * @param[in] pArguments The #IotDemoArguments_t to validate.
- *
- * @return `true` if every member of the #IotDemoArguments_t is valid; `false`
- * otherwise.
- */
-static bool _validateArguments( const IotDemoArguments_t * pArguments )
-{
-    /* Declare a status variable for this function. */
-    IOT_FUNCTION_ENTRY( bool, true );
-
-    /* Check that a server was set. */
-    if( ( pArguments->pHostName == NULL ) ||
-        ( strlen( pArguments->pHostName ) == 0 ) )
-    {
-        IotLogError( "MQTT server not set. Exiting." );
-
-        IOT_SET_AND_GOTO_CLEANUP( false );
-    }
-
-    /* Check that a server port was set. */
-    if( pArguments->port == 0 )
-    {
-        IotLogError( "MQTT server port not set. Exiting." );
-
-        IOT_SET_AND_GOTO_CLEANUP( false );
-    }
-
-    /* Check credentials for a secured connection. */
-    if( pArguments->securedConnection == true )
-    {
-        /* Check that a root CA path was set. */
-        if( ( pArguments->pRootCaPath == NULL ) ||
-            ( strlen( pArguments->pRootCaPath ) == 0 ) )
-        {
-            IotLogError( "Root CA path not set. Exiting." );
-
-            IOT_SET_AND_GOTO_CLEANUP( false );
-        }
-
-        /* Check that a client certificate path was set. */
-        if( ( pArguments->pClientCertPath == NULL ) ||
-            ( strlen( pArguments->pClientCertPath ) == 0 ) )
-        {
-            IotLogError( "Client certificate path not set. Exiting." );
-
-            IOT_SET_AND_GOTO_CLEANUP( false );
-        }
-
-        /* Check that a client certificate private key was set. */
-        if( ( pArguments->pPrivateKeyPath == NULL ) ||
-            ( strlen( pArguments->pPrivateKeyPath ) == 0 ) )
-        {
-            IotLogError( "Client certificate private key not set. Exiting." );
-
-            IOT_SET_AND_GOTO_CLEANUP( false );
-        }
-    }
-    else
-    {
-        if( pArguments->awsIotMqttMode == true )
-        {
-            IotLogError( "AWS IoT does not support unsecured connections." );
-
-            IOT_SET_AND_GOTO_CLEANUP( false );
-        }
-    }
-
-    /* No cleanup is required for this function. */
-    IOT_FUNCTION_EXIT_NO_CLEANUP();
-}
-
-/*-----------------------------------------------------------*/
-
 int main( int argc,
           char ** argv )
 {
@@ -230,16 +103,10 @@ int main( int argc,
     IotNetworkCredentials_t credentials = IOT_DEMO_CREDENTIALS_INITIALIZER,
                             * pCredentials = NULL;
 
-    /* Load the default demo arguments from the demo config header. */
-    _setDefaultArguments( &demoArguments );
-
-    /* Parse any command line arguments. */
-    IotDemo_ParseArguments( argc,
-                            argv,
-                            &demoArguments );
-
-    /* Validate arguments. */
-    if( _validateArguments( &demoArguments ) == false )
+    /* Parse and validate any command line arguments. */
+    if( IotDemo_ParseArguments( argc,
+                                argv,
+                                &demoArguments ) == false )
     {
         IOT_SET_AND_GOTO_CLEANUP( EXIT_FAILURE );
     }
