--- demos/app/win32/iot_demo_arguments_win32.c
@@ -1,43 +1,0 @@
-/*
- * Copyright (C) 2019 Amazon.com, Inc. or its affiliates.  All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-/**
- * @file iot_demo_arguments_win32.c
- * @brief Implements a function for retrieving command line arguments on Win32
- * systems.
- */
-
-/* The config header is always included first. */
-#include "iot_config.h"
-
-/* Common demo includes. */
-#include "iot_demo_arguments.h"
-#include "iot_demo_logging.h"
-
-/*-----------------------------------------------------------*/
-
-void IotDemo_ParseArguments( int argc,
-                             char ** argv,
-                             IotDemoArguments_t * pArguments )
-{
-}
-
-/*-----------------------------------------------------------*/
