+++ demos/include/iot_demo_arguments.h
@@ -87,8 +87,10 @@ typedef struct IotDemoArguments
  * @param[in] argc The argument count originally passed to main().
  * @param[in] argv The argument vector originally passed to main().
  * @param[out] pArguments Set to the arguments parsed from the command line.
+ *
+ * @return `true` if all arguments are valid; `false` otherwise.
  */
-void IotDemo_ParseArguments( int argc,
+bool IotDemo_ParseArguments( int argc,
                              char ** argv,
                              IotDemoArguments_t * pArguments );
